Flatten observations in the policy if needed
The current behavior of garage is for samplers,
algorithms, and users to flatten their observations
before feeding them into policies. This is because
garage policies flatten their observation spaces
and expect observations from that flattened observation
space.

The proposed change is to flatten observations inside of
the policies if needed. Observations should be flattened
only if they arent numpy arrays, because numpy arrays are
the type of observation that would be returned by a flattened
observation box space.

This makes ddpg-fetch-her work, however some additional changes
were needed to make it work. For example, the timelimit terminated
fetch environment only inserts timelimit truncated signals
if the environment has reached a timelimit. This doesn't work
when considering trajectory batches, which require that the
same number of keys are in every observation in a trajectory.
To get around this, if there is no time limit termination,
a time limit termination False signal is inserted.